Fix if previous_method is None
Small fix. We added that in the method_switching the key None represents the initial method. But now we have to exclude that key when we're iterating through the method_switching scheme to find which method to switch to.

Additionally, the method_history is initialized to [self.initial_method, self.initial_method] so if the method has to be switched in the first step it has a possible switch from the default switching scheme