ui method change
Took most of the functionality of method.py in pypesto into ui.py in petab_select to make petab_select more independent of the calibration tool.
The outcome is that the pipeline changed from:
-> new candidates -> calibrate -> find best ...
to now being:
-> best of last calibrated candidates -> new candidates -> calibrate ...

TODO:
Since candidates() in ui.py is changed, all functions/classes that use this function have to be changed accordingly. (to provide the previous_predecessor_model and full history, and to except to receive new candidates, history, and local history)